+++ tests/test_plugin_setup.py
@@ -26,6 +26,9 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
         settings['signer.reviewers_group'] = '{bucket_id}-{collection_id}-reviewers'
         settings['signer.alice.reviewers_group'] = 'revoyeurs'
         settings['signer.alice.source.to_review_enabled'] = 'true'
+
+        settings['signer.stage.normandy.to_review_enabled'] = 'false'
+        settings['signer.stage.normandy.group_check_enabled'] = 'true'
         return settings
 
     def test_capability_is_exposed(self):
@@ -70,6 +73,15 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
                 "destination": {"bucket": "prod",
                                 "collection": None},
                 "reviewers_group": "stage-{collection_id}-reviewers",
+            }, {
+                "source": {"bucket": "stage",
+                           "collection": "normandy"},
+                "preview": {"bucket": "preview",
+                            "collection": "normandy"},
+                "destination": {"bucket": "prod",
+                                "collection": "normandy"},
+                "reviewers_group": "stage-normandy-reviewers",
+                "group_check_enabled": True,
             }]
         }
         self.assertEqual(expected, capabilities['signer'])
@@ -216,14 +228,16 @@ class OnCollectionChangedTest(unittest.TestCase):
 
     def test_nothing_happens_when_resource_is_not_configured(self):
         evt = mock.MagicMock(payload={"action": "update", "bucket_id": "a", "collection_id": "b"})
-        sign_collection_data(evt, resources=utils.parse_resources("c/d -> e/f"))
+        sign_collection_data(evt, resources=utils.parse_resources("c/d -> e/f"),
+                             to_review_enabled=True)
         assert not self.updater_mocked.called
 
     def test_nothing_happens_when_status_is_not_to_sign(self):
         evt = mock.MagicMock(payload={"action": "update", "bucket_id": "a", "collection_id": "b"},
                              impacted_records=[{
                                  "new": {"id": "b", "status": "signed"}}])
-        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"))
+        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"),
+                             to_review_enabled=True)
         assert not self.updater_mocked.sign_and_update_destination.called
 
     def test_updater_is_called_when_resource_and_status_matches(self):
@@ -236,7 +250,8 @@ class OnCollectionChangedTest(unittest.TestCase):
             "/buckets/a/collections/b": mock.sentinel.signer
         }
         evt.request.route_path.return_value = "/v1/buckets/a/collections/b"
-        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"))
+        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"),
+                             to_review_enabled=True)
         self.updater_mocked.assert_called_with(
             signer=mock.sentinel.signer,
             storage=mock.sentinel.storage,
@@ -257,13 +272,15 @@ class OnCollectionChangedTest(unittest.TestCase):
             "/buckets/a/collections/b": mock.sentinel.signer
         }
         evt.request.route_path.return_value = "/v1/buckets/a/collections/b"
-        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"))
+        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"),
+                             to_review_enabled=True)
         assert evt.request._attachment_auto_save is True
 
     def test_updater_does_not_fail_when_payload_is_inconsistent(self):
         # This happens with events on default bucket for kinto < 3.3
         evt = mock.MagicMock(payload={"action": "update", "subpath": "collections/boom"})
-        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"))
+        sign_collection_data(evt, resources=utils.parse_resources("a/b -> c/d"),
+                             to_review_enabled=True)
 
 
 class BatchTest(BaseWebTest, unittest.TestCase):
@@ -346,6 +363,7 @@ class SourceCollectionDeletion(BaseWebTest, unittest.TestCase):
     @classmethod
     def get_app_settings(cls, extras=None):
         settings = super(cls, SourceCollectionDeletion).get_app_settings(extras)
+        settings['signer.to_review_enabled'] = 'true'
         settings['signer.stage.editors_group'] = 'something'
         return settings